requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
requests
python-multipart
pillow
tensorflow-cpu #avoid if no gpu as it consumes more spaceuvicorn main:app --reload

python-dotenv
pytest
